Bug fixes and cleanups.

+++ include/floats/Random.h
@@ -8,7 +8,7 @@ namespace fl {
 auto randomInt() -> I32 {
   std::time_t t;
   std::time(&t);
-  std::srand(t);
+  std::srand((unsigned int)t);
   return std::rand();
 }
